Split plots and plot_pipeline modules
